extract {gmp}	R Documentation

Extract or Replace Parts of a 'bigz' or 'bigq' Object

Description

Operators acting on vectors, arrays and lists to extract or replace subsets.

Usage

## S3 method for class 'bigz'
x[i=NULL, j=NULL, drop = TRUE]
## S3 method for class 'bigq'
x[i=NULL, j=NULL, drop = TRUE]

##___ In the following, only the bigq method is mentioned (but 'bigz' is "the same"): ___
## S3 method for class 'bigq'
c(..., recursive = FALSE)
## S3 method for class 'bigq'
rep(x, times=1, length.out=NA, each=1, ...)

Arguments

x
R object of class "bigz" or "bigq", respectively.

...
further arguments, notably for c().

i,j
indices, see standard R subsetting and subassignment.

drop
logical, unused here, i.e., matrix subsetting always returns a matrix, here!

times, length.out, each
integer; typically only one is specified; for more see rep (standard R, package base).

recursive
from c()'s default method; disregarded here

Examples

  a <- as.bigz(123)
  ## indexing "outside" --> extends the vectors (filling with NA)
  a[2] <- a[1]
  a[4] <- -4

  ## create a vector of 3 a
  c(a,a,a)

  ## repeate a 5 times
  rep(a,5)

  ## with matrix: 3 x 2
  m <- matrix.bigz(1:6,3)

  m[1,] # the first row
  m[1,, drop=TRUE] # the same: drop does *not* drop
  m[1]
  m[-c(2,3),]
  m[-c(2,3)]
  m[c(TRUE,FALSE,FALSE)]

  ##_modification on matrix
  m[2,-1] <- 11

[Package gmp version 0.7-5.1 Index]