gcdex {gmp}	R Documentation

Compute Bezoult Coefficient

Description

Compute g,s,t as as + bt = g = gcd(a,b). s and t are also known as Bezoult coefficients.

Usage

gcdex(a, b)

Arguments

a,b
either integer, numeric, character string, or of class "bigz"; If a string, either starting with "0x" for hexadecimal, "0b" for binary or without prefix for decimal values.

Value

a class "bigz" vector of length 3 with (long integer) values g, s, t.

Author(s)

Antoine Lucas

References

The GNU MP Library, see http://gmplib.org

See Also

gcd.bigz

Examples

gcdex(342,654)

[Package gmp version 0.5-14 Index]